Частично реализован тяговый привод!

+++ addons/ep20/ep20/include/kmb2.h
@@ -31,7 +31,7 @@ public:
     double getVelocityLevel();
 
     // Получить реверсивное состояние
-    float getReverseState();
+    int getReverseState();
 
     // Получить состояние разварота Ключ-карты для анимации
     double getTurn();
